submission: July 31 progress report — paper, figures, and source refresh
- Fig 3 per-cube stages regenerated on a fresh CVT/RVD run (was QEM fallback)
- scale figure re-rendered with atlas-group colours from canonical_best;
  new component-coloured PHerc1203 cube panel
- LS-PAVA described as the primary metric solve, active-set QP as comparison;
  tabu search formalized as move/acceptance lists; Huber citation added
- ribbon-coverage figure removed; results section stated affirmatively
- render_mesh.py --vcolors; Fig 3 panel renderer scripts; updated root PDF

+++ src/common/csr.c
@@ -45,7 +45,7 @@ CSR_T CSR_from_faces(Arena_T arena, const int32_t *faces, size_t nf,
     assert(faces);
 
     /* Pass 1: count degree (overcount: 2 neighbors per face vertex) */
-    int32_t *degree = (int32_t *)ARENA_CALLOC(arena, (long)nv,
+    int32_t *degree = (int32_t *)ARENA_CALLOC(arena, (size_t)nv,
                                                (long)sizeof(int32_t));
 
     for (size_t f = 0; f < nf; f++) {
